pyproject.toml
[tool.poetry]
name = "pdf-to-mp3-gui"
version = "0.1.0"
description = ""
authors = ["Shubh A Chudasama <shubh.cs@proton.me>"]
readme = "README.md"
license = "BSD 3-Clause License"

[tool.poetry.dependencies]
python = "^3.8.3"
gtts = "^2.5.1"
pyttsx3 = "^2.90"
pillow = "^10.2.0"
pdfminer-six = "^20231228"
pymupdf = "^1.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
